Group `color` text examples
